requirements.txt
requests
pandas
python-dotenv
requests
psycopg2-binary
